requirements/common.txt
django-configurations==2.0
django-dotenv==1.4.1
dj-database-url==0.4.1

django-cms==3.4.1
djangocms-text-ckeditor==3.3.1
django-filer==1.2.5
cmsplugin-filer==1.1.3
django-reversion==1.10.2

solid-i18n==1.3.0
django-compressor==2.1
django-htmlmin==0.10.0
django-rosetta==0.7.12
django-admin-shortcuts==1.2.7
django-parler==1.6.5
